requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp
azure-functions
azure-mgmt-automation
azure-identity
azure-mgmt-resource
git+https://github.tools.sap/hec-azure/azpoe-common.git#egg=azpoe-common